backend/requirements.txt
requests==2.26.0
django==3.2.8
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.0
django-filter==21.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
asgiref==3.5.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
Pillow==8.3.1
webcolors==1.12